Rename StaticCOntentCrawler to AbstractContentCrawler and related file name changes.
