+++ src/crawlee/parsel_crawler/_parsel_crawling_context.py
@@ -4,7 +4,7 @@ from parsel import Selector
 from typing_extensions import Self
 
 from crawlee._utils.docs import docs_group
-from crawlee.static_content_crawler._static_crawling_context import ParsedHttpCrawlingContext
+from crawlee.abstract_http_crawler._http_crawling_context import ParsedHttpCrawlingContext
 
 
 @dataclass(frozen=True)
