--- src/crawlee/static_content_crawler/__init__.py
@@ -1,11 +1,0 @@
-from ._static_content_crawler import StaticContentCrawler, StaticContentCrawlerOptions
-from ._static_content_parser import BlockedInfo, StaticContentParser
-from ._static_crawling_context import ParsedHttpCrawlingContext
-
-__all__ = [
-    'BlockedInfo',
-    'ParsedHttpCrawlingContext',
-    'StaticContentCrawler',
-    'StaticContentCrawlerOptions',
-    'StaticContentParser',
-]
